Fixed an issue regarding a NullReferenceException
Using a template with an empty (null) value for Author.Url caused the code to basicly call (null).ToString() which lead to throwing a NullReferenceException.
This can be fixed by using the ?.-operator instead of the normal .-operator before calling ToString(). This way the call will just return null if the input is null and a string if its not.