Updated Mentions to work,
Needs discussion. In short the Parse and User Ids are mutually exclusive.
I have added the ability to supply no ID and that will be considered as "Parse all users", and the ability to supply an ID which will be "Parse this specific user".

According to the documentation, having both Parse Users and Specified Users is a validation error and will throw 400: Bad Request. The solution I used was to always use Parse Users when available and ignoring the Specified Users, but this maybe not desired outcome? Please discuss.