Mentionables (#533)
* Initial Mentions and DiscordMentions

The DiscordMentions will be for internal user and the Mentions interface will be exposed on message functions.

* Added Mentions to the payload

* Sealed interface

* Added exception

* Added CreateMessage

* Added mentions to UploadFileAsync

* Moved to Create Payload

Edits don't effect pings
- Emzi

* Added mentions to webhooks

They didn't have the mentions and now they do. They are added as optionals for backwards compatability.

* Fixed build failure.

* Added test command

d#mention will trigger some user mentions for testing.

* Updated Mentions to work,

Needs discussion. In short the Parse and User Ids are mutually exclusive.
I have added the ability to supply no ID and that will be considered as "Parse all users", and the ability to supply an ID which will be "Parse this specific user".

According to the documentation, having both Parse Users and Specified Users is a validation error and will throw 400: Bad Request. The solution I used was to always use Parse Users when available and ignoring the Specified Users, but this maybe not desired outcome? Please discuss.

* Updated Documentation

I confused myself with the Everyone mention. I assumed it would allow me to mention everyone I listed but it is only specifically for the @ everyone and @ here.

I rewrote the documentation to be more clear about all the fields.

* Added All static.

Maxine suggested to add Everyone static. I believe All is more appropriately suited though.

Unsure about EveryoneMention.All however. I implemented it for consistency with the others but it isn't required.