Add overloads of reaction methods with string
You only need to know the unicode representation (or the name and ID) of an emoji/guild emoji to be able to use it in reactions. Giving the user the option to use a string instead of the more complex DiscordEmoji object increases flexibility, versatility and elegance. I believe this feature was clearly missing from the library, as it appears it was even intended for these overloads to exist.