Single Line Pages
# Summary
This proposed fix prevents the first line being added becoming a page by itself. 

# Changes Done
Added a check that ensures the current line is after the cut off point before splitting it.

# Discussion
Is this the intended behaviour? 
Currently, the first line will always be split as a separate page. Its unclear if this is intentional and causes confusion while trying to use this function.