Fixed VoiceNext ignoring incoming packets after a while.
RTP sends each voice packet with a 16-bit unsigned sequence number so that we can ensure proper ordering.  This sequence number wraps around to zero after hitting 65535, and clients are expected to be able to deal with this.

RFC 3550, A.1, suggests an algorithm for handling this.  The full version of that algorithm would require more pervasive changes (for example, it suggests a very short queue to handle minor inversions, whereas we just send everything along immediately and drop out-of-order arrivals).

Rather than going all-in on those ideas, for now we've just changed how we handle sequence numbers that get close to either end of the range, under the assumption that very low and very high sequence numbers appearing close together is an indication that a wraparound occurred recently, and so we should be treating the very low ones as *higher* than the very high ones during the transition.  Of course, once we're comfortably past a "very low" threshold for a transition, we have to start looking for another transition again.

Resolves #1084