fix(commands): Tell Discord the max value of parameters, even if the user doesn't specify
This commit fixes an issue wherein integer types could/would silently overflow if provided an input larger than the datatype can represent.

i.e., Discord only has two number types `INTEGER` and `NUMBER`, the latter being decimal. Both support numbers up to 2^53, but the C# types range from 2^8 (byte) to 2^64 (long); in the case where 2^53 < 2^N, this is fine, since generally the minimum value isn't also in a range wherein this is a problem.

The issue occurs when 2^53 > 2^N, as users may provide values greater than what can be represented in N bits, which overflows, and the converter returns an empty value, causing execution to silently fail.

This is fixed by hard-coding the max values for supported integer types, while taking the smaller number fi specified, as to not break the MinMaxValue attribute.