Changed ConfigureEvents to ConfigureEventHandlers in the preamble docs
When I was busy setting up my bot I realized that I had an error after copying "The finished product", when I dug a little deeper I found that the DiscordClientBuilder did not contain a definition for "ConfigureEvents". In the "help-and-support" section in the Discord me and the user "plerx0715" agreed that this is an error in the documentation.