feat(commands): Add interface for resolving prefixes
This adds a new interface (IPrefixResolver) that allows for resolving prefixes for text commands.

Implementations of this interface are registered with the IoC container, and thus can participate in IoC - this allows implementers to, for instance, allow resolving delegates on a per-guild basis, allowing for dynamic prefixes in a more ergonomic manner.