fix(d*spatch): Removing JSON from JObject may be problematic.
This commit removes the removal of JSON data from the raw payload. At the time of the original commit, this was intended to be a performance optimization to prevent re-deserrializing the same data (guilds) twice.

Similar efforts have been made in the past, but apparently have proven to be problematic with regards to how JSON.NET appears to handle this removal for reasons unbeknownst to me.

Nevertheless, the removal of this should not bring any meaningful downsides, and should in theory fix long-standing issues with certain bots that relied on these IDs being passed, which ultimately were not, due to this serialization issue.