fix(core): Restart heartbeat task when resuming to prevent race condition
This commit fixes (hopefully) a race condition when we attempt to resume. Previously, the heartbeat task continued to run even when the transport service was disconnected (which disposes of the underlying socket)

This aims to rectify that issue by stopping the heartbeat task via cancelling the CTS, which is consistent with our other reconnection methods.